Action Refactoring Phase 1
In this phase, we make sure that the success and failure of actions can only be defined
for similar types of actions i.e. ListViewActions can be chained together or

However, we can allow a more restrictive action call as in the case of ListViewAction
calling a ViewAction by wrapping a ViewAction in a ListViewAction. The reverse is however
not possible i.e. ViewAction cannot call a ListViewAction.

We'll do Phase 2 only if needed i.e. allow any arbitrary chaining e.g. ListItemAction can
call ViewAction on ListView rather than just on ListViewItem.